requirements.txt
# --- Web Framework ---
fastapi==0.115.14
uvicorn==0.34.3
starlette==0.46.2

# --- Database ---
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.2

# --- Authentication ---
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.5.0

# --- Utils ---
python-dotenv==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
typing_extensions==4.14.0
typing-inspection==0.4.1

# --- Required by other packages ---
anyio==4.9.0
h11==0.16.0
click==8.2.1
colorama==0.4.6
greenlet==3.2.3
Mako==1.3.10
MarkupSafe==3.0.2
six==1.17.0
rsa==4.9.1
pyasn1==0.6.1
ecdsa==0.19.1
idna==3.10
sniffio==1.3.1
annotated-types==0.7.0
